fix(ci): line-limit guardrail — extract remote-branch diff helpers
remote_repair.py reached 1501 lines after the task_tag commit work, one
line over the 1500-line first-party maintainability limit. Move the
cohesive remote-branch changed-path diff cluster
(_changed_paths_since_remote_branch, _remote_branch_diff_base_and_changed_paths,
_merge_base_with_head, _changed_paths_between_ref_and_head,
_protected_file_diffs_for_status_paths) into a new sibling
remote_repair_diffs.py and rewire the runner mixin, dropping remote_repair.py
to 1353 lines. Behavior is unchanged — methods are still attached to
PullRequestMonitorRunner via the delegate mixin.
